Add more information on output
- min required OSNR
- min propagated GSNR
- PDL, PMD, CD penalties

This enables user to diagnose the case whan GSNR is OK but path is failing

change the test_parser:
- keep the testTopology_response.json file as it is as input file for
test_csv_response_generation so that previous json exports are still tested.
- use a new testTopology_response_expected.json to check the actual json
generation with the additionnal informations
- add a 'fake response.json file with various types of response to better
test the jsontocsv function

Add more info in the logger for the case of no_ feasible_mode and add a test
for this case.

refactor a bit the functions

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: I92105f58adb7303f3d1475e4b21bd3e36e227090